US7 profile integrado con back, eliminacion app productManagerApp
