+++ PetsyApp/petsy/urls.py
@@ -11,5 +11,11 @@ urlpatterns = [
     url(r'^logout/$', views.logout_user, name="logout"),
     url(r'^create/$', views.create, name="create"),
     url(r'^profile/$', views.profile, name="profile"),
-    url(r'^shop/$', views.shop, name="shop")
+    url(r'^shop/(?P<id_shop>.*)$', views.shop, name="shop"),
+
+    url(r'^product/(?P<id_product>.*)/$', views.get_product_by_id, name='product_by_id'),
+    url(r'^product/$', views.create_product, name='create_product'),
+    url(r'^products/$', views.index, name='index'),
+    url(r'^review_product_by_id/$', views.review_product_by_id, name='review_product_by_id'),
+    url(r'^profile/$', views.get_user, name='get_user')
 ]
