--- PetsyApp/productManagerApp/__init__.py
@@ -1,1 +1,0 @@
-from django.forms.forms import *  # NOQA
