--- PetsyApp/productManagerApp/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-from .models import *
-
-admin.site.register(Shop)
-admin.site.register(Review)
-admin.site.register(Product)
