--- PetsyApp/productManagerApp/urls.py
@@ -1,11 +1,0 @@
-from django.conf.urls import url
-
-from . import views
-
-urlpatterns = [
-    url(r'^(?P<id_product>.*)/$', views.get_product_by_id, name='product_by_id'),
-    url(r'^$', views.create_product, name='create_product'),
-    url(r'^products/$', views.index, name='index'),
-    url(r'^review_product_by_id/$', views.review_product_by_id, name='review_product_by_id'),
-    url(r'^user/$', views.get_user, name='get_user')
-]
